+++ source/lysis_mode/LysisWorld.h
@@ -42,7 +42,7 @@ public:
    * Definitions of setup functions, expanded in LysisWorldSetup.cc
    */
   void SetupHosts(long unsigned int* POP_SIZE);
-  void SetupSymbionts(int* total_syms);
+  void SetupSymbionts(long unsigned int* total_syms);
 
 
   /**
