+++ source/pgg_mode/PGGWorld.h
@@ -31,7 +31,7 @@ public:
   * Definitions of setup functions, expanded in PGGWorldSetup.cc
   */
   void SetupHosts(long unsigned int* POP_SIZE);
-  void SetupSymbionts(int* total_syms);
+  void SetupSymbionts(long unsigned int* total_syms);
 
 
   /**
